First attempt of normalization

+++ models/data.go
@@ -3,7 +3,7 @@ package models
 type Data struct {
 	Country   string
 	Year      int
-	Waste     int
+	Waste     float64
 	Latitude  string
 	Longitude string
 }
